Test Google Cloud Platform, Readjust size of login/register button and check if firebase authentication and firestore works as expected.
