+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "29.0.2"
     defaultConfig {
         applicationId "com.dragonnetwork.hihealth"
-        minSdkVersion 27
+        minSdkVersion 26
         targetSdkVersion 29
         versionCode 1
         versionName "1.0"
