+++ app/src/main/AndroidManifest.xml
@@ -9,18 +9,20 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".welcome.LoginActivity"></activity>
-        <activity android:name=".MainActivity">
+        <activity
+            android:name=".user.LoginActivity"
+            android:label="@string/title_activity_register">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <activity android:name=".MainActivity">
+
+        </activity>
         <activity android:name=".user.UserProfile"></activity>
-        <activity
-            android:name=".user.LoginActivity"
-            android:label="@string/title_activity_login"></activity>
         <activity
             android:name=".user.SignupActivity"
             android:label="@string/title_activity_register" />
