+++ app/src/main/java/com/dragonnetwork/hihealth/MedicationAdaptor.java
@@ -9,6 +9,8 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.dragonnetwork.hihealth.data.Medication;
+
 import java.util.List;
